add alert dialog component, add alert box when delete plan

+++ src/app/plan-list/plan-list.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { Plan } from '../Plan';
 import { PlanService } from '../plan.service';
 import { ActivatedRoute } from '@angular/router';
+import { MatDialog, MatDialogConfig } from '@angular/material';
+import { MyDialogComponent } from '../my-dialog/my-dialog.component';
 
 
 @Component({
@@ -12,7 +14,9 @@ import { ActivatedRoute } from '@angular/router';
 export class PlanListComponent implements OnInit {
   plans: Plan[];
   constructor(private planservice: PlanService,
-  private activatedRoute: ActivatedRoute) { }
+  private activatedRoute: ActivatedRoute,
+  public dialog: MatDialog) { }
+
 
   ngOnInit() {
    // console.log(this.activatedRoute.snapshot.params.plan);
@@ -31,12 +35,29 @@ export class PlanListComponent implements OnInit {
   }
   
   deletePlan(id: string): void{
-
-     this.planservice.deletePlan(id).subscribe(data=>{
-       this.getPlans();
-     });
+     this.openModal(id);
   }
 
+
+  openModal(id: string) {
+    const dialogConfig = new MatDialogConfig();
+   dialogConfig.disableClose = true;
+    dialogConfig.autoFocus = true;
+    dialogConfig.data = {
+    id: 1,
+    title: "Hi there"
+    };
+   const dialogRef = this.dialog.open(MyDialogComponent, dialogConfig);
+   dialogRef.afterClosed().subscribe(result => {
+     if(result){
+        this.planservice.deletePlan(id).subscribe(data=>{
+        this.getPlans();
+      });
+     }
+    console.log("Dialog was closed!")
+    console.log(result)
+   });
+    }
 }
